+++ utils/catalog-parser/src/test/java/de/sovity/edc/utils/catalog/DspCatalogServiceTest.java
@@ -46,7 +46,7 @@ class DspCatalogServiceTest {
         var result = CompletableFuture.completedFuture(StatusResult.success(catalogJson.getBytes(StandardCharsets.UTF_8)));
         when(catalogService.requestCatalog(eq(endpoint), eq("dataspace-protocol-http"), eq(QuerySpec.max()))).thenReturn(result);
         var monitor = mock(Monitor.class);
-        var dataOfferBuilder = new DspDataOfferBuilder(new TitaniumJsonLd(monitor), monitor);
+        var dataOfferBuilder = new DspDataOfferBuilder(new TitaniumJsonLd(monitor));
 
         return new DspCatalogService(catalogService, dataOfferBuilder);
     }
